hueristic/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# hueristic/heuristic.py: 5
Pillow == 7.1.1

# hueristic/heuristic.py: 12,13,14
colormath == 3.0.0

# hueristic/heuristic.py: 3,4,6
matplotlib == 3.2.1

# hueristic/heuristic.py: 2
numpy == 1.18.2

# hueristic/heuristic.py: 1
opencv_python == 4.2.0.32

# hueristic/heuristic.py: 8
pandas == 1.0.3

# hueristic/heuristic.py: 7,9,10
scipy == 1.4.1
